+++ types/number/mode.d.ts
@@ -1,0 +1,5 @@
+declare enum NumberMode {
+    Unit4 = 2019,
+    Unit = 2020
+}
+export default NumberMode;