+++ types/region/index.d.ts
@@ -1,0 +1,3 @@
+import RegionKr from 'region-name-kr/region-kr';
+export default class Region extends RegionKr {
+}