+++ types/text/dateMode.d.ts
@@ -1,0 +1,12 @@
+declare enum dateMode {
+    TRIM = "",
+    SPACE = " ",
+    SECOND = 7,
+    MINUTES = 6,
+    HOUR = 5,
+    DATE = 3,
+    MONTH = 2,
+    YEAR = 1,
+    MONTH_DATE = 4
+}
+export default dateMode;