+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 # 체인지 로그
+## 0.5.1
+`.d.ts` 파일 추가
+
 ## 0.5.0
 * `hangul-js`의 함수 추가
 * `correctJosa` 함수 추가
+
 ## 0.4.1
 * `dateMode.MONTH_DATE`가 작동하지 아니하는 오류 수정
+
 ## 0.4.0
-* Date 관련 지원 추가(dateTime, date, time)
+* Date 관련 지원 추가(`dateTime`, `date`, `time`)
+
 ## 0.3.0
-* Number 업데이트(toNumber)
+* Number 업데이트(`toNumber`)
+
 ## 0.2.0
-* Number 지원(toKorean)
+* Number 지원(`toKorean`)
+
 ## 0.1.0
-* Region 지원(region-name-kr)
+* Region 지원(`region-name-kr`)